Move encoding into a separate package
